requirements.txt
# AgenticOS Python Dependencies
# Install: pip install -r requirements.txt

# Required
PyYAML>=6.0.0,<7.0.0

# Optional (enables log schema validation in doctor --strict)
jsonschema>=4.0.0,<5.0.0
